+++ features/maps/maps.html
@@ -98,3 +98,5 @@
 </html>
 
 <script src="./mapsRenderer.js"></script>
+<script src="./konva.min.js"></script>
+<script src="./mapCanvasRenderer.js"></script>
